Now scripts don't fail if SERVER_PID isn't running

+++ tests/server/test_ende.sh
@@ -4,7 +4,10 @@
 set -e
 
 clean_up() {
-    kill $SERVER_PID
+    if ps -p $SERVER_PID > /dev/null
+    then
+         kill $SERVER_PID
+    fi
 }
 trap clean_up EXIT
